restructure shared/events for Email, OX and LDAP domains
