+++ src/shared/events/email/disabled-email-address-generated.event.ts
@@ -1,5 +1,5 @@
-import { BaseEvent } from './base-event.js';
-import { EmailAddressID, PersonID, PersonReferrer } from '../types/index.js';
+import { EmailAddressID, PersonID, PersonReferrer } from '../../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
 
 /**
  * This event should be triggered when a new email-address is generated for a user, DIRECTLY DISABLED and persisted successfully in the database.
