+++ src/shared/events/email/email-address-already-exists.event.ts
@@ -1,4 +1,4 @@
-import { BaseEvent } from './base-event.js';
+import { BaseEvent } from '../base-event.js';
 
 export class EmailAddressAlreadyExistsEvent extends BaseEvent {
     public constructor(
