+++ src/shared/events/email/email-address-deleted-in-database.event.ts
@@ -1,7 +1,7 @@
-import { EmailAddressID, PersonID } from '../types/aggregate-ids.types.js';
-import { BaseEvent } from './base-event.js';
-import { EmailAddressStatus } from '../../modules/email/domain/email-address.js';
-import { OXUserID } from '../types/ox-ids.types.js';
+import { EmailAddressID, PersonID } from '../../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
+import { EmailAddressStatus } from '../../../modules/email/domain/email-address.js';
+import { OXUserID } from '../../types/ox-ids.types.js';
 
 export class EmailAddressDeletedInDatabaseEvent extends BaseEvent {
     public constructor(
