+++ src/shared/events/email/email-address-deleted.event.ts
@@ -1,7 +1,7 @@
-import { EmailAddressID, PersonID, PersonReferrer } from '../types/aggregate-ids.types.js';
-import { BaseEvent } from './base-event.js';
-import { EmailAddressStatus } from '../../modules/email/domain/email-address.js';
-import { OXUserID } from '../types/ox-ids.types.js';
+import { EmailAddressID, PersonID, PersonReferrer } from '../../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
+import { EmailAddressStatus } from '../../../modules/email/domain/email-address.js';
+import { OXUserID } from '../../types/ox-ids.types.js';
 
 export class EmailAddressDeletedEvent extends BaseEvent {
     public constructor(
