+++ src/shared/events/email/email-address-disabled.event.ts
@@ -1,4 +1,4 @@
-import { BaseEvent } from './base-event.js';
+import { BaseEvent } from '../base-event.js';
 
 export class EmailAddressDisabledEvent extends BaseEvent {
     public constructor(
