+++ src/shared/events/email/email-addresses-purged.event.ts
@@ -1,6 +1,6 @@
-import { PersonID, PersonReferrer } from '../types/aggregate-ids.types.js';
-import { BaseEvent } from './base-event.js';
-import { OXUserID } from '../types/ox-ids.types.js';
+import { PersonID, PersonReferrer } from '../../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
+import { OXUserID } from '../../types/ox-ids.types.js';
 
 /**
  * Published when all EmailAddresses for a Person were removed.
