+++ src/shared/events/email/kafka-email-address-changed.event.spec.ts
