+++ src/shared/events/email/kafka-email-address-changed.event.ts
@@ -1,5 +1,5 @@
 import { EmailAddressChangedEvent } from './email-address-changed.event.js';
-import { KafkaEvent } from './kafka-event.js';
+import { KafkaEvent } from '../kafka-event.js';
 
 export class KafkaEmailAddressChangedEvent extends EmailAddressChangedEvent implements KafkaEvent {
     public get kafkaKeyPersonId(): string {
