+++ src/shared/events/email/kafka-email-address-deleted.event.ts
@@ -1,4 +1,4 @@
-import { KafkaEvent } from './kafka-event.js';
+import { KafkaEvent } from '../kafka-event.js';
 import { EmailAddressDeletedEvent } from './email-address-deleted.event.js';
 
 export class KafkaEmailAddressDeletedEvent extends EmailAddressDeletedEvent implements KafkaEvent {
