+++ src/shared/events/email/kafka-email-address-generated.event.spec.ts
