+++ src/shared/events/email/kafka-email-address-generated.event.ts
@@ -1,5 +1,5 @@
 import { EmailAddressGeneratedEvent } from './email-address-generated.event.js';
-import { KafkaEvent } from './kafka-event.js';
+import { KafkaEvent } from '../kafka-event.js';
 
 export class KafkaEmailAddressGeneratedEvent extends EmailAddressGeneratedEvent implements KafkaEvent {
     public get kafkaKeyPersonId(): string {
