+++ src/shared/events/email/kafka-email-addresses-purged.event.ts
@@ -1,4 +1,4 @@
-import { KafkaEvent } from './kafka-event.js';
+import { KafkaEvent } from '../kafka-event.js';
 import { EmailAddressesPurgedEvent } from './email-addresses-purged.event.js';
 
 export class KafkaEmailAddressesPurgedEvent extends EmailAddressesPurgedEvent implements KafkaEvent {
