+++ src/shared/events/ldap/ldap-email-address-deleted.event.ts
@@ -1,5 +1,5 @@
-import { BaseEvent } from './base-event.js';
-import { PersonReferrer } from '../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
+import { PersonReferrer } from '../../types/aggregate-ids.types.js';
 
 export class LdapEmailAddressDeletedEvent extends BaseEvent {
     public constructor(
