+++ src/shared/events/ldap/ldap-entry-deleted.event.ts
@@ -1,5 +1,5 @@
-import { BaseEvent } from './base-event.js';
-import { PersonID, PersonReferrer } from '../types/index.js';
+import { BaseEvent } from '../base-event.js';
+import { PersonID, PersonReferrer } from '../../types/aggregate-ids.types.js';
 
 export class LdapEntryDeletedEvent extends BaseEvent {
     public constructor(
