+++ src/shared/events/ox/ox-account-deleted.event.ts
@@ -1,6 +1,6 @@
-import { BaseEvent } from './base-event.js';
-import { OXUserID } from '../types/ox-ids.types.js';
-import { PersonReferrer } from '../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
+import { OXUserID } from '../../types/ox-ids.types.js';
+import { PersonReferrer } from '../../types/aggregate-ids.types.js';
 
 export class OxAccountDeletedEvent extends BaseEvent {
     public constructor(
