+++ src/shared/events/ox/ox-email-address-deleted.event.ts
@@ -1,6 +1,6 @@
-import { BaseEvent } from './base-event.js';
-import { OXContextID, OXContextName, OXUserID } from '../types/ox-ids.types.js';
-import { PersonReferrer } from '../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
+import { OXContextID, OXContextName, OXUserID } from '../../types/ox-ids.types.js';
+import { PersonReferrer } from '../../types/aggregate-ids.types.js';
 
 export class OxEmailAddressDeletedEvent extends BaseEvent {
     public constructor(
