+++ src/shared/events/ox/ox-metadata-in-keycloak-changed.event.ts
@@ -1,6 +1,6 @@
-import { BaseEvent } from './base-event.js';
-import { OXContextName, OXUserID, OXUserName } from '../types/ox-ids.types.js';
-import { PersonID } from '../types/aggregate-ids.types.js';
+import { BaseEvent } from '../base-event.js';
+import { OXContextName, OXUserID, OXUserName } from '../../types/ox-ids.types.js';
+import { PersonID } from '../../types/aggregate-ids.types.js';
 
 /**
  * Thrown when ID_OX has been filled in KC for a user after successfully creating OX-account for the user.
